requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==2.2.4
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
numpy==1.17.0
pandas==0.25.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyMySQL==0.9.3
pytesseract==0.2.7
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.6
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
